更改金额相关字段类型为Long

+++ uctp-modules/uctp-business/src/main/java/com/newtouch/uctp/module/business/controller/app/account/cash/vo/CashDetailRespVO.java
@@ -36,7 +36,7 @@ public class CashDetailRespVO {
     private String tranRecordNo;
 
     @Schema(description = "支付金额")
-    private Integer payAmount;
+    private Long payAmount;
 
     @Schema(description = "交易类型")
     private String tradeType;
@@ -54,7 +54,7 @@ public class CashDetailRespVO {
     private String contractNo;
 
     @Schema(description = "账户余额")
-    private Integer accountBalance;
+    private Long accountBalance;
 
     @Schema(description = "提现状态")
     private String presentState;
